Rebasing after burner
